chore: better folder structure and code structure
