+++ constants/populateDB.js
@@ -1,0 +1,16 @@
+const colors = ["Violet", "Indigo", "Blue", "Green", "Yellow", "Orange", "Red"]
+const sizes = ["XS", "S", "M", "L", "XL"]
+const stores = [
+    "Chic Boutique",
+    "Style Avenue",
+    "Glamour Attic",
+    "Fashionista Finds",
+    "Trendy Threads",
+    "Couture Corner",
+    "Elegant Edge",
+    "Vogue Vault",
+    "Dapper Den",
+]
+const styles = ["A-line", "Shift", "Wrap", "Sheath", "Fit and Flare", "Maxi", "Midi", "Mini"]
+
+module.exports =  { colors, sizes, stores, styles }
